Nhato - "Let Me Drive" (Lua)
Player field only in motion for the appropriate times now, and a few
dark/no dark mods applied for emphasis. Road actually drives.

Looks like I need to separate this code into BG/FG changes... sucks for
the random sine spline! but maybe it won't be so bad in the end, the
trees may not need to be so super realistic?