Nhato - "Let Me Drive" (Lua)
OK!! As far as I'm concerned, we've reached the bare minimum of
completeness - of course there are still some things I'd like to add,
but it could ship tonight.
Still need to resolve BGBrightness...it's a PREFSMAN setting so it might
need to be done ahead of time instead of through the file.